Fixed password verification and hashing
Verification wasn't working because the PHP hashing algorithm is different to the SQL hashing algorithm.

To fix this, the INSERT statements had to include a PHP hashed password, not an SQL hashed password when a new user is inserted into the DB.

The register_process.php page follows this new method of hashing and verification.